Sizes of data types

+++ 0x00-hello_world/6-size.c
@@ -16,6 +16,6 @@ int main(void)
 	printf("Size of an int:%i.\n", (unsigned int)sizeof(i));
 	printf("Size of a long int:%d.\n", (unsigned long)sizeof(j));
 	printf("Size of a long long int:%lu.\n", (unsigned long)sizeof(k));
-	printf("Size of a float:%f.\n", (unsigned long)sizeof(f));
+	printf("Size of a float:%d.\n", sizeof(f));
 	return (0);
 }
